+++ src/routes/routes.js
@@ -1,6 +1,7 @@
 import {Route} from 'react-router-dom';
 import Store from '../components/store/store';
-import {CreateCustomer} from '../components/CreateCustomer';
+import {CreateCustomer} from '../components/Customer/CreateCustomer';
+import {GetToken} from '../components/Customer/GetToken'
 import Cart from '../components/Cart/Cart';
 
 
@@ -10,6 +11,7 @@ const routes = () => {
             <Route path="/store" component={Store} />
             <Route path="/createcustomer" component={CreateCustomer} />
             <Route path="/cart" component={Cart} />
+            <Route path="/gettoken" component={GetToken} />
         </>
     )
